Changed implementation of validator
